Added new flight logic
When a descending of more than 200 meters is detected and the chemistry
experiment has not been done yet, it is initiated regardless of being at
a lower height than we'd like.
Some improvements which were not yet in the repo.
Images are taken at a 5sec interval instead of the previous 15sec.